milestone-3/requirements.txt
streamlit
opencv-python
numpy
Pillow
matplotlib
fpdf
streamlit-image-comparison
